[Remove] Drawing component, button for swiching on/of drawing [Add] new drawing method which use Line tag from konva
